An example to demonstrate "Controller-Payload-Target" interaction
The example demonstrates the between the controller module, payload modules
and the target module. See README for details.

A target module is the module to be analysed.

Payload modules contain at least the replacement functions that implement
the logic of the operations: which data to collect when the corresponding
functions are called by the target module, fail the call or not during
fault simulation, etc.

A controller module maintains the list of curently registered payload
modules, detects loading of the target module and instruments the latter
using the replacement tables provided by the payload modules.