Prepared the system to generate the HTML docs from DocBook sources
By default, the docs are not regenerated when 'make' or 'make doc' are called.
But they are if -DKEDR_GENERATE_DOC:BOOL=TRUE is specified for cmake at
the configuration stage. See sources/doc/Readme.txt for details.

'make install' now installs docs to (the stored ones or the generated ones,
depending on the configuration).