Rework indicator templates.
Now there is internal representation of the indicator model data(data-file).
This internal model is now difficult for use externally, so it is not export outside. But this model is usefull for generation indicator source file.

Also added tests for this internal representation of indicator.

Also add tests for fault simulation with NULL init and destroy functions.