leak_check: Blocks with the same addresses are now filtered
The same addresses can only appear if leak_check missed some deallocation
calls.

Only the most recently allocated block is shown in 'possible_leaks'.

If KEDR_DEBUG is defined, the module reports each block, for which a newer
block with the same address exists, to the system log. Otherwise, only
the number of such "orphaned" blocks is output (again, to the system log).