The presence of *node*-functions is now checked in a more reliable way
On some systems, the functions like kmem_cache_alloc_node_notrace() are
exported but not avaivable for the kernel modules because they are ifdef'ed
out. On other systems the functions may be available.

For such functions, building of a probe module is now used to determine if
the function can be called from a kernel module.