util: revisited the implementation of kedr_save_stack_trace()
save_stack_trace() function is not always available in the kernel, even
if reliable stack traces are supported (that is, CONFIG_STACKTRACE may
be undefined in this case too).

kedr_save_stack_trace() now works as follows.

If reliable stack traces are supported and save_stack_trace() is available,
it is used. If the function is not available but still reliable stack
traces are supported, we provide our implementation of save_stack_trace().
It is basically the original implementation copied from the kernel, with
only minor changes.

If reliable stack traces are not supported, kedr_save_stack_trace()
stores only one stack frame passed to it as one of the arguments.