Set version to 0.4-devel
It seems reasonable to change the versioning scheme a bit to clearly
indicate which version is now being prepared.

"x.y-devel" now means the development version of KEDR x.y. It will be used
before KEDR x.y is released. After that, the development version will change
to "x.y+1-devel" or the like.

In addition, the micro version should only be used when it is really needed.
"KEDR x.y" means KEDR "x.y.0".