A header with #defines for the used functions is now generated
During the configuration phase, func_def.h header file is now generated in
the top build directory. For each kernel function that is available on the
current system and is processed by at least one payload module, the file
contains the following line:
    #define KEDR_FUNC_USED_<function name>
Here "<function name>" denotes the name of the corresponding function.

This may simplify the implementation of the payload modules to some extent,
especially for the payloads that are not generated from the collection of
.data files.