LeakCheck: Added processing of kasprintf() and kvasprintf()
kasprintf() and kvasprintf() allocate memory using SLAB allocator, so it is
desirable for LeakCheck to process these functions too. Unfortunately, the
standard mechanism of calling the pre, post and replacement handlers used in
KEDR does not go well with the functions with variable argument lists like
kasprintf().

As a temporary workaround, support for kasprintf() and kvasprintf() has been
provided to LeakCheck only as a separate plugin, created in a special way.
It also does not actually use the .data files for these functions, they are
added to be able to use the standard function configuration facilities.

Support for these functions in the call monitoring and fault simulation
tools is not provided. Besides that, no other payload module may process
these functions if LeakCheck is active.

These limitations could be removed in the future versions of KEDR.