LeakCheck: Added a test to detect changes in the implementation of kmalloc
The test uses LeakCheck for a module that calls kmalloc() and alloc_pages()
to allocate several memory blocks of different size, and the corresponding
functions to free them. The size may be a compile-time constant or may be
determined in runtime only.

This should allow to detect at least some of the changes in the
implementation of kmalloc, etc. The developers seem to rename the functions
kmalloc expands to from time to time, change their parameters a bit and so
on. Although this way to detect such changes is not 100% reliable, it may
help.