By default, LeakCheck is to be used, if enabled, when KEDR is launched
When KEDR is launched without the configuration file specified, the
configuration for LeakCheck will be used. This is similar to Valgrind and
its Memcheck tool.

Before this commit, the call monitoring subsystem was loaded by default. As
LeakCheck seems to be used more often than call monitoring, it is reasonable
to change the defaults accordingly.