LeakCheck: The results can now be output to system log too
In addition to the files in debugfs, LeakCheck can now output its results to
system log. This can be useful, for example, if the system under test is
unstable or its UI becomes unusable. It is often possible to have the system
log messages to be output to a serial console and to retrieve them from
another machine.

Parameter "syslog_output" of the main LeakCheck module controls this
behaviour. By default, the output to the system log is enabled. Set
syslog_output to 0 if you would like to disable it.