Payload templates: Use va_copy in every template-generated function which call more than one other functions with va_list argument. kedr_function_support.c: replace incorrect memcpy() call with explicit NULL-assignments to structure fields.
On x86_64 it is not sufficient to obtain independed copy of 'va_list'-typed variable via common pointer copiing.
In templates. when process function with variable number of parameters, for each call from intermediate/replacement function use 'args_copy' variable, obtained via va_copy() from 'args', and pass it to other functions.
For correct processing of function which takes 'va_list'-typed argument, user should provide corresponding 'args_copy_init' and 'args_copy_destroy' parameters, which describe copiing of that argument and destroying that copy.
That parameters are not documented yet, but used in datafile for kvasprinf() function. Test is added for fault simulation in case of that parameters(also for kvasprintf function).